requirements/base.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
pymongo>=4.0.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
ta>=0.7.0
tensorflow>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0